move plugin to laser module
