+++ mythril/laser/plugin/plugins/dependency_pruner.py
@@ -1,7 +1,7 @@
 from mythril.laser.ethereum.svm import LaserEVM
-from mythril.laser.ethereum.plugins.plugin import LaserPlugin
-from mythril.laser.ethereum.plugins.signals import PluginSkipState
-from mythril.laser.ethereum.plugins.implementations.plugin_annotations import (
+from mythril.laser.plugin import LaserPlugin
+from mythril.laser.plugin.signals import PluginSkipState
+from mythril.laser.plugin.plugins.plugin_annotations import (
     DependencyAnnotation,
     WSDependencyAnnotation,
 )
