+++ mythril/laser/plugin/signals.py
